+++ test/providers/loop_provider_test.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:cs_app/models/error_detection_challenge.dart';
 import 'package:cs_app/models/loop_challenge.dart';
 import 'package:cs_app/providers/loop_provider.dart';
 import 'package:cs_app/services/database_service.dart';
@@ -16,6 +17,10 @@ class _FakeDatabaseService implements DatabaseService {
 
   @override
   Future<List<LoopChallenge>> fetchPuzzlesByType(String type) async => _puzzles;
+
+  @override
+  Future<List<ErrorDetectionChallenge>> fetchErrorDetectionPuzzles() async =>
+      const [];
 
   @override
   Stream<List<LoopChallenge>> getLoopPuzzles() => Stream.value(_puzzles);